change some detail by demand

+++ Judger_Test/Judger_Test/checkers/special_judger.py
@@ -1,40 +1,44 @@
 import lorun
 import os
-
+from consts import VerdictResult
 
 def create_special_judger(**sj_detail):
     code = sj_detail['sj_code']
     name = sj_detail['sj_name']
     code_file_path = './'+name+'.cpp'
-    log_file_path = './log.txt'
+    log_file_path = './sj_logs/'+name+'_log.txt'
     checker_exec = '../testlib/bin/' + name
+    try:
+        with open(code_file_path, 'w') as cf:
+            cf.write(code)
 
-    with open(code_file_path, 'w') as cf:
-        cf.write(code)
+        with open(log_file_path, 'w') as log_file:
+            run_cfg = {
+                'args': ['g++', code_file_path, '-o', checker_exec, '-I', '../testlib/'],
+                'fd_in': 0,
+                'fd_out': 0,
+                'fd_err': log_file.fileno(),
+                'timelimit': 5000,  # 5 S
+                'memorylimit': 251072,
+            }
+            result = lorun.run(run_cfg)
+            # print(result)
+            feedback = dict()
+            feedback['log'] = log_file.read()
+            feedback['result'] = result['result']
 
-    with open(log_file_path, 'w') as log_file:
-        run_cfg = {
-            'args': ['g++', code_file_path, '-o', checker_exec, '-I', '../testlib/'],
-            'fd_in': 0,
-            'fd_out': 0,
-            'fd_err': log_file.fileno(),
-            'timelimit': 5000,  # 5 S
-            'memorylimit': 251072,
-        }
-        result = lorun.run(run_cfg)
-        # print(result)
-        feedback = dict()
-        feedback['log'] = log_file.read()
-        feedback['result'] = result['result']
-
+            # TODO send_result_back
+            if result['result'] != VerdictResult.AC:
+                pass
+            else:
+                pass
+    except Exception:
         # TODO send_result_back
-        if result['result'] != 0:
-            pass
-        else:
-            pass
-
-    os.remove(code_file_path)
-    os.remove(log_file_path)
+        pass
+    finally:
+        if os.path.exists(code_file_path):
+            os.remove(code_file_path)
+        #os.remove(log_file_path)
 
 
 # a test
